fix(jellycompat): support macOS operation locks
